\name{aafChromosome-class}
\docType{class}
\alias{aafChromosome-class}
\title{Class aafChromosome, a class for gene chromosome assignments}
\description{
  An abstraction for gene gene chromosome assignments from Bioconductor
  data packages.
}

\section{Objects from the Class}{
Objects are generally created by the \code{\link{aafChromosome}}
constructor.  Objects can also be created manually by calls of the
form \code{new("aafChromosome", description)}.
}
\section{Slots}{
  \describe{
    \item{\code{.Data}:}{Object of class \code{character}}
  }
}
\section{Extends}{
Class \code{character}, from data part.
}

\section{Methods}{
No methods defined with class "aafChromosome" in the signature.
See generic implementations of \code{\link{getText}}, \code{\link{getURL}},
\code{\link{getHTML}}, \code{\link{getTD}}, and \code{\link{getCSS}}.
}

\author{Colin A. Smith, \email{annaffy@colinsmith.org}}

\seealso{\link{aafChromosome}}

\keyword{classes}
