\name{getCSS-methods}
\docType{methods}
\title{Methods for function getCSS}
\alias{getCSS-methods}
\alias{getCSS}
\description{
  Methods to get relevant stylesheet lines for an object.
}

\section{Methods}{\describe{
\item{object = ANY}{Returns an empty character vector.}
}}

\note{
  For information about other implementations of this method, 
  see documentation of the respective class.  
}

\seealso{
  \link{aafList-class},
  \link{aafPubMed-class},
  \link{aafGO-class},
  \link{aafPathway-class},
  \link{aafSigned-class}
}

\keyword{methods}
