\name{getTD-methods}
\docType{methods}
\title{Methods for function getTD}
\alias{getTD-methods}
\alias{getTD}
\alias{getTD,ANY-method}
\description{
  Methods to get an HTML table cell representation of an object.
}

\section{Methods}{\describe{
\item{object = ANY}{Returns <td> tag containing HTML representation of \code{object}. Sets class attribute to \code{class(object)}. }
}}

\note{
  For information about other implementations of this method, 
  see documentation of the respective class.  
}

\seealso{
  \link{aafList-class},
  \link{aafGO-class},
  \link{aafPathway-class},
  \link{aafIntensity-class}
}

\keyword{methods}
