\name{getText-methods}
\docType{methods}
\title{Methods for function getText}
\alias{getText-methods}
\alias{getText}
\alias{getText,ANY-method}
\description{
  Methods to get a textual representation of an object.
}

\section{Methods}{\describe{
\item{object = ANY}{Returns a comma delimeted list of the elements in \code{list}.}
}}

\note{
  For information about other implementations of this method, 
  see documentation of the respective class.  
}

\seealso{
  \link{aafList-class},
  \link{aafCytoband-class},
  \link{aafGO-class},
  \link{aafGOItem-class},
  \link{aafPathway-class},
  \link{aafPathwayItem-class}
}

\keyword{methods}
