\name{getURL-methods}
\docType{methods}
\title{Methods for function getURL}
\alias{getURL-methods}
\alias{getURL}
\description{
  Methods to get a URL link to a web resource for an object.
}

\section{Methods}{\describe{
\item{object = ANY}{Returns an empty character vector.}
}}

\note{
  For information about other implementations of this method, 
  see documentation of the respective class.  
}

\seealso{
  \link{aafList-class},
  \link{aafGenBank-class},
  \link{aafLocusLink-class},
  \link{aafCytoband-class},
  \link{aafUniGene-class},
  \link{aafPubMed-class},
  \link{aafGO-class},
  \link{aafGOItem-class},
  \link{aafPathwayItem-class}
}

\keyword{methods}
